requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
diff-match-patch==20241021
Django==5.2.7
django-cors-headers==4.9.0
django-cuser==2017.3.16
django-filter==25.2
django-import-export==4.3.12
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
modeltranslation==0.25
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
